chore(recommend): hand the data pipeline over to tracks B and C
A 가 만든 것을 A 없이 계속 돌릴 수 있게 정리했습니다 (A-13).

- `feature_version` 을 `seeds/` 내용 해시에서 끌어냅니다 (`v1-15a8c5`).
  `"v1"` 로 굳혀 두면 검수를 반영한 2회차 피처가 1회차와 같은 판 번호를 달고
  나가서, B·C 가 캐시한 옛 값과 새 값을 구분할 방법이 없습니다. 에러가 아니라
  점수만 조금씩 어긋나므로 아무도 알아채지 못합니다. 사람이 기억해서 올리는
  규칙으로 두지 않고 사전에서 끌어내, 한 줄만 고쳐도 값이 저절로 바뀝니다.
- `make help` 의 이름 칸을 14 에서 18 로 넓혔습니다. `normalize-batch`(15자)·
  `popularity-build`(16자)처럼 긴 이름이 설명과 붙어 읽히지 않았습니다.
- README 2절에 재실행 한 줄을 적었습니다 — 레시피가 늘거나 사전이 바뀌면
  `make renormalize` 하나면 됩니다.
- 골든 픽스처를 다시 뽑았습니다. A-12 이전에 만들어 `cluster_id` 가 30건 전부
  null 이었는데, 이제 실제 군집 번호가 들어갑니다. B·C 가 이 파일을 계약으로
  쓰므로 비워 두면 재정렬을 못 짭니다. `stats_version` 도 5에서 6이 됐습니다.
- `cluster_build.py` 의 중복 `int()` 를 지웠습니다 (ruff RUF046). A-12 때
  파일 단위로만 검사해 놓친 것입니다.

확인한 것

- `recipe_feature` 46,353건 · `flavor_vec` 46,353 · `cluster_id` 46,353 (군집 50개)
- `make feature-test` 체크 8개 전부 초록 · `make data-gate` 전부 통과
- ruff · mypy 63파일 · pytest 206건 통과
- 사전을 한 줄 고치면 지문이 바뀌고 되돌리면 원래 값으로 돌아오는 것 확인

+++ src/features/recommend/ingest/cluster_build.py
@@ -225,7 +225,7 @@ def build(n_ingredients: int = 536) -> ClusterStats:
 
     st.recipes, st.assigned, st.n_clusters, st.max_share = load_cluster_stats()
     # 고르게 퍼진 10개를 봅니다. 앞에서 10개만 보면 한쪽에 쏠린 것을 놓칩니다.
-    picked = sorted({int(round(i * (K - 1) / 9)) for i in range(10)})
+    picked = sorted({round(i * (K - 1) / 9) for i in range(10)})
     for cid, title in load_cluster_samples(picked):
         st.samples.setdefault(cid, []).append(title)
     return st
